regex: remove insensitive flag
I don't know if this flag was there on purpose but when trying the
regex in the example, I noticed `bug-123` was also modified.

I guess that's better to let the user pick what it wants. Here I
suggest to remove the `i` flag and the user can use `(BUG|bug)` to have
the same behaviour. But if needed, we could also add a config option.